Repackage helper classes to avoid conflict with Bedrock invoke API instrumentation
