+++ instrumentation/aws-bedrock-runtime-2.26.25/src/main/java/llm/converse/models/ModelResponse.java
@@ -5,7 +5,7 @@
  *
  */
 
-package llm.models;
+package llm.converse.models;
 
 public interface ModelResponse {
     // Operation types
