+++ instrumentation/aws-bedrock-runtime-2.26.25/src/main/java/llm/converse/models/converse/ConverseModelRequest.java
@@ -5,12 +5,12 @@
  *
  */
 
-package llm.models.converse;
+package llm.converse.models.converse;
 
 import com.newrelic.api.agent.NewRelic;
-import llm.models.ModelRequest;
+import llm.converse.models.ModelRequest;
 import software.amazon.awssdk.services.bedrockruntime.model.ContentBlock;
-import software.amazon.awssdk.services.bedrockruntime.model.ConverseStreamRequest;
+import software.amazon.awssdk.services.bedrockruntime.model.ConverseRequest;
 import software.amazon.awssdk.services.bedrockruntime.model.InferenceConfiguration;
 import software.amazon.awssdk.services.bedrockruntime.model.Message;
 
@@ -18,35 +18,28 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 
-// TODO nothing in this class has been tested as streaming support isn't yet implemented. It seems that
-//  ConverseStreamRequest provides the same APIs as ConverseRequest, so perhaps this class isn't necessary
-//  and ConverseModelRequest can just handle both request types (ConverseStreamRequest and ConverseRequest).
-//  Might just need to modify the constructor of ConverseModelRequest to accept a more generic
-//  shared interface/abstract class (e.g. BedrockRuntimeRequest, AwsRequest, or SdkRequest) and
-//  just cast to the appropriate request type if need be.
-
 /**
  * Stores the required info from the Bedrock ConverseRequest.
  * Avoids holding a reference to the request object to prevent potential memory issues.
  */
-public class ConverseModelStreamRequest implements ModelRequest {
+public class ConverseModelRequest implements ModelRequest {
     private String modelId = "";
     private Integer maxTokens = 0;
     private Float temperature = 0.0F;
     private int numOfMessages = 0;
     private List<Message> messages = new ArrayList<>();
 
-    public ConverseModelStreamRequest(ConverseStreamRequest converseStreamRequest) {
-        if (converseStreamRequest != null) {
-            modelId = converseStreamRequest.modelId();
-            InferenceConfiguration inferenceConfiguration = converseStreamRequest.inferenceConfig();
+    public ConverseModelRequest(ConverseRequest converseRequest) {
+        if (converseRequest != null) {
+            modelId = converseRequest.modelId();
+            InferenceConfiguration inferenceConfiguration = converseRequest.inferenceConfig();
 
             if (inferenceConfiguration != null) {
                 maxTokens = inferenceConfiguration.maxTokens();
                 temperature = inferenceConfiguration.temperature();
             }
-            if (converseStreamRequest.hasMessages()) {
-                messages = converseStreamRequest.messages();
+            if (converseRequest.hasMessages()) {
+                messages = converseRequest.messages();
                 numOfMessages = messages.size();
             }
         } else {
