+++ instrumentation/aws-bedrock-runtime-2.26.25/src/test/java/software/amazon/awssdk/services/bedrockruntime/BedrockRuntimeClient_InstrumentationTest.java
@@ -25,15 +25,14 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
-import static llm.events.LlmEvent.LLM_CHAT_COMPLETION_MESSAGE;
-import static llm.events.LlmEvent.LLM_CHAT_COMPLETION_SUMMARY;
-import static llm.models.TestUtil.REQUEST_CONTENT_TEXT;
-import static llm.models.TestUtil.REQUEST_MODEL_ID;
-import static llm.models.TestUtil.RESPONSE_CONTENT_TEXT;
-import static llm.models.TestUtil.STOP_REASON;
-import static llm.models.TestUtil.assertErrorEvent;
-import static llm.models.TestUtil.assertLlmChatCompletionMessageAttributes;
-import static llm.models.TestUtil.assertLlmChatCompletionSummaryAttributes;
+import static llm.converse.events.LlmEvent.LLM_CHAT_COMPLETION_MESSAGE;
+import static llm.converse.events.LlmEvent.LLM_CHAT_COMPLETION_SUMMARY;
+import static llm.converse.models.TestUtil.REQUEST_CONTENT_TEXT;
+import static llm.converse.models.TestUtil.REQUEST_MODEL_ID;
+import static llm.converse.models.TestUtil.RESPONSE_CONTENT_TEXT;
+import static llm.converse.models.TestUtil.STOP_REASON;
+import static llm.converse.models.TestUtil.assertLlmChatCompletionMessageAttributes;
+import static llm.converse.models.TestUtil.assertLlmChatCompletionSummaryAttributes;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
